feat(ingest): JDF CI default parse path with OMP staging and structured confidence
- pdf_to_parse_bundle: canonical parse entrypoint returning jdf, chunks,
  text, real page_count, parser/source metadata, parse/OCR confidence
  (None when unreported, 0.0 passes through), and tables/images/figures
  as distinct first-class payloads with counts and asset_summary
- import_project_pdf + jdf ingest + substrate ingest: JDF CI first with
  logged best-effort fallbacks, OMP parse artifact staged immediately
  after parse with explicit confidence kwargs and vault-row linkage
- substrate_vault migration v26: parser_name, source_kind, confidence
  (REAL-or-NULL), asset counts, asset_summary, omp_artifact_id columns;
  save/upsert/list persist and return them
- omp/models: float confidence, parser/source metadata and structured
  assets carried in ParseArtifactPayload and artifact payload/provenance
  with is-not-None guards (unknown stays None, 0.0 survives)
- confidence.py: rewr
- pdf_to_parse_bundle: canonical parse entrypoint returning jdf, chunks,
  text, real page_count, parser/source metadatabre  text, real page_count, parser/source metadata, parse/OCR confidence
 se  (None when unreported, 0.0 passes through), and tables/images/figuen  as distinct first-class payloads with counts and asset_summary
- imp-b- import_project_pdf + jdf ingest + substrate ingest: JDF CI fis:  logged best-effort fallbacks, OMugh/asset retention/fallback
  coverag  after parse with explicits.py and test_confidence.py

+++ prompt_matrix/models/jdf.py
@@ -215,9 +215,10 @@ class JDFDocumentTree(BaseModel):
     truth_ledger: dict[str, float | str | int] = Field(default_factory=dict)
     body: list[JDFSectionNode] = Field(default_factory=list)
 
-    # Confidence fields for verification results
-    confidence: int | None = None
-    confidenceBreakdown: dict[str, int] | None = None
+    # Confidence fields for verification results — float, not int: scores are
+    # ratios and an int forces every caller to round before it can store.
+    confidence: float | None = None
+    confidenceBreakdown: dict[str, float] | None = None
     lowConfidenceNodes: list[dict[str, Any]] | None = None
 
     # OMP lineage fields for audit trail
